Add BaseExchange class
